Update deployment hooks

+++ src/lib/operations/remove-deployment.js
@@ -1,19 +1,27 @@
 import { getKubernetesClient } from 'kitegg-directus-extension-common'
 import k8s from '@kubernetes/client-node'
 
-export async function removeDeployment(deploymentObject, res = undefined) {
-  const appsClient = getKubernetesClient(
-    deploymentObject.namespace,
-    k8s.AppsV1Api
+export async function removeDeployment(id, res = undefined) {
+  const appsClient = getKubernetesClient(undefined, k8s.AppsV1Api)
+  const { body: existing } = await appsClient.listDeploymentForAllNamespaces(
+    undefined,
+    undefined,
+    undefined,
+    `llp.kitegg.de/objectId=${id}`,
+    1
   )
-  await appsClient.deleteNamespacedDeployment(
-    deploymentObject.name,
-    deploymentObject.namespace,
-    undefined,
-    undefined,
-    undefined,
-    undefined,
-    'Background'
-  )
-  return { deleted: deploymentObject.id }
+  if (existing.items.length > 0) {
+    for (const item of existing.items) {
+      await appsClient.deleteNamespacedDeployment(
+        item.metadata.name,
+        item.metadata.namespace,
+        undefined,
+        undefined,
+        undefined,
+        undefined,
+        'Background'
+      )
+    }
+  }
+  return { deleted: id }
 }
